refactor component to be functional rather than class
